Add Fixup && Add Images
